+++ mala/descriptors/acelib/clebsch_gordan_coupling.py
@@ -1,6 +1,6 @@
 """CG coupling functions."""
 
-from mala.descriptors.acelib import ace_coupling_utils as ace_coupling_utils
+from mala.descriptors.acelib import coupling_utils as ace_coupling_utils
 
 
 def get_coupling(Clebsch_Gordan, ldict, L_R=0, use_permutations=True):
